OMAP3LOGIC: Modified board detect routine runtime placement
The board detect routine doesn't work if the USB driver has had
any initialization prior to running.  At that point, it'll depend
on the connection of an OTG device to determine which board the
system is running on.  Fixed this by placing the detect routine
earlier in the system startup.  Kept the printout in a similar
placement to overall booting.

Signed-off-by: Tim Nordell <tim.nordell@logicpd.com>