OMAP3LOGIC: Modified the timings for the ISP1760 once again
Some of the timing parameters were set in a way that didn't matter due to
the value of RDCYLCETIME (essentially it would deassert those signals
anyways).  Changed these timings a bit, and moved the setting from
BUSTURNAROUND into CYCLE2CYCLEDELAY with the same chip select
enabled instead.  This way, the extra delay required between accesses
won't affect other pending accesses to other chips as much (i.e. such
as NAND).

Technically, the settings are a bit conservative as 1 clock cycle
is roughly 12ns.  The read and write parameters probably could be
tightened a little bit if _really_ needed.  The Cycle2CycleDelay is a
couple of cycles higher than what seems to be the minimum "working"
setting.

Signed-off-by: Tim Nordell <tim.nordell@logicpd.com>